pyproject.toml
[tool.poetry]
name = "beranabyte-api"
version = "0.1.0"
description = "API for beranabyte.com website"
authors = ["Kalkidan Betre <kalbetre.dev@outlook.com>"]

packages = [
    { include = "api" }
]

[tool.poetry.scripts]
start = "api.app:start_app"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
motor = "^2.5.1"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.2"
pydantic = "^1.8.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.90.0"}
websockets = "^10.1"
email-validator = "^1.1.3"
firebase-admin = "^5.1.0"
requests = "^2.26.0"
python-multipart = "^0.0.5"
geocoder = "^1.38.1"
user-agents = "^2.2.0"
black = "^21.12b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"
pytest-asyncio = "^0.16.0"
coverage = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
